+++ CHANGELOG.md
@@ -1,3 +1,34 @@
+## [v2.0.23](https://pro-gallery-2-0-23.surge.sh) (16/07/2020)
+ 
+#### GALLERY
+ -  SSR improve: item dimension
+
+---
+## [v2.0.22](https://pro-gallery-2-0-22.surge.sh) (16/07/2020)
+ 
+#### GALLERY
+ -  export blueprints in the lib
+ 
+---
+## [v2.0.21](https://pro-gallery-2-0-21.surge.sh) (16/07/2020)
+ 
+ -  blueprint manager is no longer a singleton
+ - reverted "changed pro-gallery/lib to be only ESM."
+
+---
+## [v2.0.20](https://pro-gallery-2-0-20.surge.sh) (15/07/2020)
+ 
+#### GALLERY
+ -  changed pro-gallery/lib to be only ESM.
+
+---
+## [v2.0.19](https://pro-gallery-2-0-19.surge.sh) (15/07/2020)
+ 
+#### GALLERY
+ -  render videoItemPlaceholder in videoItem
+ -  LeanGallery supported in Ssr
+
+---
 ## [v2.0.18](https://pro-gallery-2-0-18.surge.sh) (15/07/2020)
 
 #### GALLERY